requirements.txt
opencv_contrib_python==4.5.1.48
pytesseract==0.3.7
imutils==0.5.4
Keras==2.4.3
numpy==1.19.5
Pillow==8.1.2
